qtapp/requirements.txt
pillow>=9.0.0
pyqt5>=5.15.0